[fix]: <cr> preserves sentence
Previously, if <cr> is hit mid-sentence, then a new list item would be
created and the rest of the sentence would be lost. This commit fixes
the issue by checking current mode and cursor position and creates a new
list item only if we are in insert mode and at eol, else default <cr>
behaviour is invoked.